@@ -1,14 +1,16 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="NLog" Version="4.7.9" />
-    <PackageReference Include="System.Waf.Wpf" Version="6.3.0" />
+    <PackageReference Include="System.Waf.Wpf" Version="6.4.0" />
   </ItemGroup>
 
   <ItemGroup>
@@ -16,11 +18,11 @@
   </ItemGroup>
 
   <ItemGroup>
+    <None Update="NLog.config" CopyToOutputDirectory="PreserveNewest" />
+
     <Compile Include="..\NLogCodeSample\NLogHelper.cs" Link="NLogHelper.cs" />
     <Compile Include="..\NLogCodeSample\MainWindow.xaml.cs" Link="MainWindow.xaml.cs" />
     <Page Include="..\NLogCodeSample\MainWindow.xaml" Link="MainWindow.xaml" Generator="MSBuild:Compile" />
-
-    <None Update="NLog.config" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
 
 </Project>
